Add option to hide file extensions during import
When you add media clips to your project, the default setting is for the
clips to receive the name of the file as the clip name. For example, if
you add 'Scenes/011/11A-3.MOV' into your project, the name of the clip
will be '11A-3.MOV'.

This commit adds an optional user preference to omit the file extension
on the clip name when adding files into your project. For example, the
previously mentioned clip would now receive the name '11A-3' during
import instead.

This preference is controlled by a checkbox under:
    Edit -> Preferences -> Editing ->
        [ ] Hide file extensions when importing Clips

For backwards compatibility, this option is disabled by default, so
existing users should not notice any change in behavior.