Round FPS to whole numbers for NTSC time code
This update fixes time code calculations for NTSC frame rates like
23.976 and 29.97.

Before this update, a 60 second (24 * 60 frame) clip in 23.976 frame
rate would display a length of "00:01:00:01" in the media clip view
list. And a timeline made up of 60 instances of this same clip would
end on "01:00:03:13" as the last frame.

After this update, the same clip shows a length of "00:01:00:00" in the
media clip view list, and a timeline made up of 60 instances of this
same clip ends on "00:59:59:23".