Filenames in Media-List - issue 267
Preference setting (In View tab) to show full file names in the media list.
When unticked (Default) it will show as it is now, when ticked all files will show their full file name.
No need to restart the application. Setting will be activated once a file is added or removed in the media list.