Unpickling...¿? #341
Removed print of project.name in the debugging message. This will allow the project to load if it has non-latin characters in the name.
No real reason found why it is a problem, but any manipulation on project.name (If it contains non-latin characters, like characters with tilde, accents, cedilles etc.) causes the project load to stop.