filters: add frei0r defish0r
While it may not be suitable for all lenscorrection, in many ways it's
superior to the other lenscorrection filters.

As it offers autoscaling and proper pixel interpolation.

I've had success using it to remove distortion from a non-fisheye lens
as well, and I would encourage anybody to try defish0r before trying
the other lenscorrection plugins.