Add audio sync fix when timeline has compound clips
Prior to this commit, exportardour.py did not take into account
compound clips. This means that if a compound clip appeared on
the timeline, it got completely ignored, but it also skewed the
timeline for all subsequent clips in the same playlist. Basically
like perfoming a delete edit on the compound clip.

Actually extracting audio from compound clips seems non-trivial.
Even some of the well-known proprietary programs had limitations
on OMF/AAF export of audio in compound clips. So this change
doesn't add the ability to extract audio from compound clips.

What this change does do is to try and recognize compound clips
when they appear in the MLT XML file, and respect the length that
their in and out points add to the overall duration of the
playlist. Or in other words, the compound clips still won't come
through, but at least all of the other non-compound clips that
show up after it in the playlist will still be in sync.

A warning is printed to stderr when a compound clip is detected.