Add 'Flowblade' program name to Ardour project
The Ardour session file format has two fields for the name of the
program that created and modified the Ardour session.

Previously, exportardour.py was using "exportardour.py" as the
program name. However, now that it is being integrated into
Flowblade, I have changed this value to "Flowblade".

What I really wanted to do was to reference entries in
appconstants.py for the Flowblade program name and version, and
concatenate them together. However, I didn't find any suitable
fields in there, so I hard-coded it for now.