Remove left over Python 2 string decode() calls
I tried to render a movie using custom render args from Flowblade
on the master branch. However, it bumped into a decode() call on
a string.

This would have worked on Python 2. But decode() is no longer a
valid method on strings in Python 3.

There were two main instances of this:

1) A key and value in renderconsumer.py, which were previously
   being checked to ensure they were in ASCII format

2) A decode() call on a string in gmic.py

I removed all of these decode() calls in this commit, but I'm not
completely sure that everything is correct.

For one thing, it is now possible to have non-ASCII characters
get through in renderconsumer.py. I'm not sure if that's an issue
or not. This commit just removes those decode() calls, but if they
still need to be ASCII, then something else should be put in place
of the code that was removed. On the plus side, it was not possible
to render at all with custom render args before, and it works now.

The gmic.py change was not tested. However, I did look at the
decode() call, and it looked like it was referencing a string.
However, I would strongly encourage testing this assumption!