Fix atomicfile and reintroduce os.rename()
Previously, issue #459 was resolved by replacing the os.rename()
call in AtomicFileWriter with an shutil.move() call. However,
the real underlying problem was elsewhere, and shutil.move()
cannot always offer the same guarantees that os.rename() can offer
on a single filesystem with standard POSIX semantics.

I looked into this, and there was a bug in the previous
implementation. When temp files were created, they were not always
guaranteed to be on the same filesystem as the final destination
file.

This commit fixes that situation, by explicitly using the final
destination directory as the basis for the temp file it creates.
Since the temp file and the destination file are now guaranteed to
be in the same directory, rename() is just renaming a file inside
of one directory, which won't span multiple filesystems.