Improve Ardour export integration
tools/exportardour.py started out as a command-line script with
no dependencies on Flowblade. But now it's integrated into Flowblade,
and can take advantage of more of Flowblade's features.

Since the command-line interface was basically a proof of concept, the
command-line interface is dropped here, in favor of making Ardour
session exports perform better from within the Flowblade GUI.

- All CLI/GUI condition code is gone now.
- atomicfile is used when writing the Ardour session file
- The Flowblade version number is included in the created/modified
  fields of the exported Ardour session file
- MLT XML playlist and Flowblade track name mapping now works when
  an alternate number of audio/video tracks is used in the Flowblade
  sequence being exported
- The Flowblade copyright header was added to the top of the file