Save project using atomicfile (in Python 3)
The Flowblade project file uses the Python pickle format.

In Python 3, strings and bytes must be kept separate.

The pickle format is binary.

When writing a pickle file, you must open the filehandle with the
"wb" mode (write binary).

AtomicFileWriter was previously only using "w" mode, even when "wb"
was specified.

Now, AtomicFileWriter respects the file write mode setting when
opening files. As a result, saving project files can be done
atomically again, and once again uses atomicfile as a wrapper
around the file writes to avoid the possibility of the project
file being corrupted if the save doesn't complete.