Fix bin count updates when no bin is selected
The projectaction module has a function called
update_current_bin_files_count() that is called whenever anything
is added or removed from a bin. The purpose of this function is to
update the item count in the selected bin, to keep the count up
to date.

This function works by figuring out which bin is selected, and
then updating that bin. But what happens if no bin is selected?

It is possible to deselect a bin, without selecting another one.
One way to do this is to control click on the current bin in the
Bins panel. If you do this, adding and deleting clips from the bin
will not update the count, because there is no selected bin.

This commit adds another branch to the
update_current_bin_files_count() function, to handle the case where
no bin is selected. Previously, this branch returned early. Now,
if this condition occurs, the counts for all of the bins in the
project will be updated.