Fix '+' to zoom in timeline hot key
Update the flowblade.xml key shortcut map, to fix the use of the
'+' key to zoom in on the timeline.

Previously, plus was configured in this file, but in practice you
can't press '+' without also pressing SHIFT, and so this key
combination was never mapped.

This commit adds a SHIFT modifier to the plus hotkey. It also adds
another mapping of the '=' key as an alias for zoom in, so that
you can zoom the timeline in and out with the two keys side by side,
instead of having to push shift and then + to zoom in.