Add selectable sample rate to Ardour export
Previously, all Ardour exports defaulted to a 48 kHz sampling rate. This is
usually the right choice for a film or video project, but there are other
sample rates available, and some projects might want to choose a different
one.

This commit adds a sample rate combo select box to the Ardour export
dialog. It still defaults to 48 kHz as before. But now it also offers
three other choices.

The sample rates now available are:

  44.1 kHz
  48 kHz (default)
  88.2 kHz
  96 kHz

Some audio interfaces go up even higher than that, to 192 or 384. I've
never heard of a film or video project using them though, and question
the diminishing returns. It would be easy to add more sample rates
later if it really turns out to be useful to anyone.