Major improvements to USB handling
The previous implementation of the USB HID subsystem used the pyusb library,
in synchronous mode, which polled repeatedly for input.

This commit is a major refactoring of the USB subsystem, which brings
several substantial improvements.

The pyusb dependency has been completely eliminated. Instead, USB connections
are now implemented in terms of libusb 1.0, which was the underlying library
that pyusb itself was using. The resulting code is a bit more verbose, but
also provides a lot more control and flexibility. Most notably, the USB HID
subsystem now performs asynchronous reads of the USB device, which is a
major performance improvement. This was not possible with the pyusb library.

In the previous implementation, a GTK timeout handler was triggered every
5 milliseconds to perform a USB read, with a timeout of 1 millisecond. This
actually worked reasonably well, but if you go look at the USB traffic with
Wireshark, it was very chatty and was actually talking to the device that
often.

In the current implementation, a GTK timeout handler is triggered every 10
milliseconds, which simply goes and checks whether any asynchronous USB
transfers are ready to be read. This does not block or timeout, and
happens instantly. Meanwhile, during setup, 10 USB transfers were set up
to be in flight, being passed back and forth to send and receive messages,
and events only occur when the device actually has new input to send.
These transfers have no timeouts, so the chattiness is kept to a minimum.
Having a queue with multiple transfers in flight also ensures that we capture
every single update from the device, which improves the perception of update
latency, and also improves reliability, especially with the shuttle.

Finally, the "endpoint_in" and "endpoint_out" XML tags in the USB HID driver
config files now use the hex address of the endpoints, instead of numbers
describing which endpoint it is counting from zero. The XML config files
have been updated accordingly. Similarly, the "alternate" field has been
removed, as libusb does not require that field, and neither do any of the
devices we're likely to talk to.