Fixed typo and renaming of parameters: Glitch
The "RGB Shift" filter parameter has a typo "Verical" corrected to "Vertical".
Since "Shift" is a key on the keyboard, it is not translated into other languages, which means that translation into any other language of this parameter will not be available. I suggest the names of the "Glitch" filter parameters from Kdenlive.